requirements.txt
requests==2.28.1
python-dateutil==2.8.2
pandas==1.5.0
python-dotenv==1.0.0
